updated dependencies

+++ src/py_layer.cpp
@@ -62,7 +62,7 @@ void python_export_layer(py::module &m) {
         .def_property_readonly("transmissionBottomTop", [](const LayerMode &m) -> MatrixX { return m.transmissionBottomTop; }, D(LayerMode, transmissionBottomTop));
 
     py::class_<Layer>(m, "Layer", D(Layer))
-        .def(py::init<const VectorX &, const VectorX &, int>(), py::arg("nodes"), py::arg("weights"), py::arg("nFourierOrders") = 1)
+        .def(py::init<const VectorX &, const VectorX &, size_t>(), py::arg("nodes"), py::arg("weights"), py::arg("nFourierOrders") = 1)
         .def(py::init<const Layer&>())
         .def("reverse", &Layer::reverse, D(Layer, reverse))
         .def("clear", &Layer::clear, D(Layer, clear))
